Implement chmod, chown, and utimens.
Store uid, gid, and mode in LocalGridFile. These get stored in the "owner", "group" (both names, not uid/gid), and "mode" keys.
It's a shame that mongo::GridFS/GridFile don't provide a way to add/access your own keys in the file object.
Modified the LGF setters and getters, sorry.